Refactor re-bootstrapping
